Fix bug in sum of products
When using the sum of products algorithm from https://eprint.iacr.org/2022/367.pdf there is a restriciton for canonical inputs that 2(p - 1)^2 < p*R, where R = 2^(64*n) for elements represented by n limbs.

This commit includes a prime of this form into the tests and refactors the sum of products method to gracefully handle this case.